Refaktorer, flytt ut funksjoner og fjern companion objects
